Organisation des routes dans les différents controllers - Organisation des repertoires des pages et des translations
