+++ src/Web/MainBundle/Resources/translations/CTI/Informatique/bureautique.fr.yml
