+++ src/Web/MainBundle/Resources/translations/CTI/Informatique/infographie.fr.yml
