+++ src/Web/MainBundle/Resources/translations/CTI/Informatique/informatique.en.yml
