+++ src/Web/MainBundle/Resources/translations/CTI/Vacances/vacances.fr.yml
